refactor: extract crt, pdfwriter and object into separate files (modules)
